+++ src/raytracer/accelerator/aabb.h
@@ -10,12 +10,14 @@ namespace raytracer
 	{
 	public:
 		Aabb() = default;
+
 		Aabb(const point3& a, const point3& b);
 
 		[[nodiscard]] point3 minimum() const;
+
 		[[nodiscard]] point3 maximum() const;
 
-		[[nodiscard]] bool hit(const Ray& r, double t_min, double t_max) const;
+		[[nodiscard]] bool hit(const Ray& r, float t_min, float t_max) const;
 
 		static Aabb surroundingBox(const Aabb& box0, const Aabb& box1);
