+++ src/raytracer/hittable/movingsphere.cpp
@@ -7,24 +7,26 @@
 namespace raytracer
 {
 
-	MovingSphere::MovingSphere(point3 cen0, point3 cen1, double _time0, double _time1, double r, const ref<Material>& m)
+	MovingSphere::MovingSphere(point3 cen0, point3 cen1, float _time0, float _time1, float r, const ref<Material>& m)
 			: m_center0(cen0), m_center1(cen1), m_time0(_time0), m_time1(_time1), m_radius(r), m_material(m)
-	{}
+	{
+	}
 
-	bool MovingSphere::hit(const Ray& r, double t_min, double t_max, hit_record& rec) const
+	bool MovingSphere::hit(const Ray& r, float t_min, float t_max, hit_record& rec) const
 	{
 		vec3 oc = r.origin() - center(r.time());
 		auto a = lengthSquared(r.direction());
 		auto half_b = glm::dot(oc, r.direction());
-		auto c = lengthSquared(oc) - m_radius*m_radius;
+		auto c = lengthSquared(oc) - m_radius * m_radius;
 
-		auto discriminant = half_b*half_b - a*c;
+		auto discriminant = half_b * half_b - a * c;
 		if (discriminant < 0) return false;
 		auto sqrtd = std::sqrt(discriminant);
 
 		// Find the nearest root that lies in the acceptable range.
 		auto root = (-half_b - sqrtd) / a;
-		if (root < t_min || t_max < root) {
+		if (root < t_min || t_max < root)
+		{
 			root = (-half_b + sqrtd) / a;
 			if (root < t_min || t_max < root)
 				return false;
@@ -39,13 +41,13 @@ namespace raytracer
 		return true;
 	}
 
-	point3 MovingSphere::center(double time) const
+	point3 MovingSphere::center(float time) const
 	{
 		// Linear interpolate?
-		return m_center0 + ((time - m_time0) / (m_time1 - m_time0))*(m_center1 - m_center0);
+		return m_center0 + ((time - m_time0) / (m_time1 - m_time0)) * (m_center1 - m_center0);
 	}
 
-	bool MovingSphere::boundingBox(double time0, double time1, Aabb& output_box) const
+	bool MovingSphere::boundingBox(float time0, float time1, Aabb& output_box) const
 	{
 		Aabb box0(
 				center(time0) - vec3(m_radius, m_radius, m_radius),
