+++ src/raytracer/materials/dielectric.h
@@ -8,16 +8,17 @@ namespace raytracer
 	class Dielectric : public Material
 	{
 	public:
-		explicit Dielectric(double index_of_refraction);
+		explicit Dielectric(float index_of_refraction);
+
 		~Dielectric() override = default;
 
 		bool scatter(const Ray& r_in, const hit_record& rec, color& attenuation, Ray& scattered) const override;
 
 	private:
-		static double reflectance(double cosine, double ref_idx);
+		static float reflectance(float cosine, float ref_idx);
 
 	private:
-		double m_ir;
+		float m_ir;
 	};
 
 } // raytracer
